Modified design_cpc_DOE function which gives number of rows of static collector field.
The PSA MATLAB code had a conditional statement to determine number of rows,
based on the quotient between total thermal energy required and thermal energy provided per row.
If the remainder is >=0.5, round up in number rows.

The strange alternative is to round down AND subtract 1 from the number of rows when remainder is <0.5.

I originally bypassed this logic to simply round up or down based on remainder; here I implement the aformentioned,
questionable logic used by PSA.